// xy10032/src/common/enums/action-type.enum.ts
export enum ActionType {
  CREATE = 'create',
  UPDATE = 'update',
  DELETE = 'delete',
  SUBMIT = 'submit',
  APPROVE = 'approve',
  REJECT = 'reject',
  CANCEL = 'cancel',
  PROCESS = 'process',
  SUCCESS = 'success',
  FAIL = 'fail',
  RETRY = 'retry',
  IMPORT = 'import',
  EXPORT = 'export',
  BATCH_CREATE = 'batch_create',
  BATCH_UPDATE = 'batch_update',
  BATCH_APPROVE = 'batch_approve',
  BATCH_REJECT = 'batch_reject',
  BATCH_CANCEL = 'batch_cancel',
}